Character_Level_Language_Model/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.4
cymem==2.0.5
docx2pdf==0.1.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
Flask==2.0.1
Flask-WTF==0.15.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
murmurhash==1.0.5
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pathy==0.6.0
pdf2image==1.16.0
Pillow==8.3.1
preshed==3.0.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
PyPDF2==1.26.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pytesseract==0.3.8
pywin32==227
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.15.0
sklearn==0.0
smart-open==5.2.0
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thinc==8.0.8
threadpoolctl==2.2.0
tqdm==4.62.2
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wrapt==1.12.1
WTForms==2.3.3
